requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.117
botocore==1.34.117
celery==5.4.0
celery-progress==0.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.7
daphne==4.1.2
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==0.63.3
django-anymail==10.3
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-health-check==3.18.2
django-hijack==3.4.5
django-htmx==1.18.0
django-otp==1.5.0
django-redis==5.4.0
django-storages==1.14.3
django-timezone-field==6.1.0
django-waffle==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.4
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.7.0
zope.interface==6.4.post2
